Sketch out a lowering to a core IR
Not sure this is the right approach, since we'd lose a lot of debug info without making this a lot more boring. Might be worth it later as the syntax grows some sugar.

Rewriting the compiler would mostly be a matter of updating `compiler::visitor`, so not as bad as it could be.

I'll just leave the branch here for now.