A simple list type
These are just wrappers around a rust `Vec<Value>`, not really what we want longer-term. We also can't use the current compilation strategy for anything with a spread operator in it, which is something I want later.

I wrote this for now so we can get things to a "write small test programs" state.