Pulled the primitive operations into a trait
The trait has default impls. Value and Object dispatch to the specific
types too. One maybe unintentional consequence here is that things like
`1.0 + -1` is an error.

The idea is _big_, and the code changes are messy. A lot of it's going
to change as it becomes clearer how this should work, so I'm not going
to worry about it too much.