Fix the repl
To fix the repl, we had to change how the compiler gets started up. I got the original design I was hoping for working, which is the compiler starts with an empty module, and then you 'push' syntax through it.

Each time you push, it reopens the main prototype and adds the new code there. It replaces the previous Op::Halt with a Op::Nop so the runtime doesn't miss anything.

The compiler also has to be able to recover from errors it sees when new code is pushed, so that it's left in a workable state after.