Basic plan for diagnostics
We'll start by making these `*Coordinator` types and passing them around and
then later we can work on integrating it into other modules in a better way.

Eventually we'll want the parser and compiler to attempt to recover so we can
have multiple diagnostics printed at once.

Some of this is very similar to a previous toy langage project, so we should
be able to jump-start the quality of error message printing once I pull in more
of that code.