Non-parser syntax errors
A lot. Before all the errors that came up when parsing were defined in parser, not in syntax. This is not ideal if we want the syntax to be able to define it's own errors.

I changed a bunch of stuff around in the parser to make this work, and updated the syntax to follow.

Lots of placeholder errors still need their Into<Diagnostic> written, but it's coming along.