Captures!
They work, at least for the one simple example I tried!

Needs a lot of polish. Probably pretty big changes to how GC types work to make this nice. Also relying on refcell and the fact that this is single-threaded to make this work -- it's not a viable solution longer term!