Something was very wrong with closures
I fixed the repl, but noticed the way end_scope worked was a mess. It couldn't work for my value-returning blocks. Moved from CloseCapture to Close to compensate. Reworked the compiler Capture type to use indexes properly, it was a weird rlox holdover.

Kind of wild these were so broken for so long and tests didn't catch it.