Clean up binding types
So now we use a `pub` to mark exports, and `var` works again, but it marks the binding as var. Compiles the same but we can compile-time check for `is_var` later to prevent reassignment, maybe? I have a feeling this is more complicated if we want to prevent things like `let x = []; var y = x; y.append(1)`. Nothing here's final.